# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Cambase.io is a project by Evercam.io in order to make it easier for\
    \ software developers to have a reliable, up to date source of model hardware\
    \ information available via a public API."
  version: "1.0"
  title: "Cambase.io"
  x-logo:
    url: "https://api.apis.guru/v2/cache/logo/https_twitter.com_evrcm_profile_image.jpeg"
  x-origin:
  - converter:
      url: "https://github.com/lucybot/api-spec-converter"
      version: "2.6.2"
    format: "swagger"
    url: "http://api.cambase.io/api-docs/api-docs.json"
    version: "1.2"
  x-preferred: true
  x-providerName: "cambase.io"
host: "api.cambase.io"
basePath: "/"
...
schemes:
- "http"
paths:
...
  /api/v1/recorders/{id}.json:
    get:
      tags:
      - "recorders"
      summary: "Fetches a single Recorder"
      operationId: "Api::V1::Recorders#show"
      parameters:
      - name: "id"
        in: "path"
        description: "Recorder ID"
        required: true
        type: "integer"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Recorders"
    put:
      tags:
      - "recorders"
      summary: "Updates an existing Recorder"
      operationId: "apiV1RecordersId_jsonPUT"
      parameters:
      - name: "id"
        in: "path"
        description: "Recorder ID"
        required: true
        type: "string"
      - name: "vendor_id"
        in: "formData"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "recorder[model]"
        in: "formData"
        description: "Model"
        required: true
        type: "string"
      - name: "recorder[name]"
        in: "formData"
        description: "Name"
        required: true
        type: "string"
      - name: "recorder[recorder_type]"
        in: "formData"
        description: "Type"
        required: true
        type: "string"
      - name: "recorder[resolution]"
        in: "formData"
        description: "Resolution"
        required: false
        type: "string"
      - name: "recorder[onvif]"
        in: "formData"
        description: "ONVIF"
        required: false
        type: "string"
      - name: "recorder[psia]"
        in: "formData"
        description: "PSIA"
        required: false
        type: "string"
      - name: "recorder[ptz]"
        in: "formData"
        description: "PTZ"
        required: false
        type: "string"
      - name: "recorder[discontinued]"
        in: "formData"
        description: "Discontinued"
        required: false
        type: "string"
      - name: "recorder[support_3rdparty]"
        in: "formData"
        description: "3rd pparty Camera Support"
        required: false
        type: "string"
      - name: "recorder[sd_card]"
        in: "formData"
        description: "SD Card"
        required: false
        type: "string"
      - name: "recorder[upnp]"
        in: "formData"
        description: "UPnP"
        required: false
        type: "string"
      - name: "recorder[hot_swap]"
        in: "formData"
        description: "Hot Swap"
        required: false
        type: "string"
      - name: "recorder[hdmi]"
        in: "formData"
        description: "HDMI Support"
        required: false
        type: "string"
      - name: "recorder[digital_io]"
        in: "formData"
        description: "Digital I/O"
        required: false
        type: "string"
      - name: "recorder[audio_in]"
        in: "formData"
        description: "Audio In"
        required: false
        type: "string"
      - name: "recorder[audio_out]"
        in: "formData"
        description: "Audio Out"
        required: false
        type: "string"
      - name: "recorder[input_channels]"
        in: "formData"
        description: "Input Channels"
        required: false
        type: "string"
      - name: "recorder[playback_channels]"
        in: "formData"
        description: "Playback Channels"
        required: false
        type: "string"
      - name: "recorder[usb]"
        in: "formData"
        description: "USB Ports"
        required: false
        type: "string"
      - name: "recorder[sdhc]"
        in: "formData"
        description: "SD Card (GB)"
        required: false
        type: "string"
      - name: "recorder[mobile_access]"
        in: "formData"
        description: "Mobile Access"
        required: false
        type: "string"
      - name: "recorder[alarms]"
        in: "formData"
        description: "Alarms"
        required: false
        type: "string"
      - name: "recorder[raid_support]"
        in: "formData"
        description: "Raid Support"
        required: false
        type: "string"
      - name: "recorder[storage]"
        in: "formData"
        description: "Internal Storage"
        required: false
        type: "string"
      - name: "recorder[additional_information]"
        in: "formData"
        description: "Additional Information"
        required: false
        type: "string"
      - name: "recorder[default_username]"
        in: "formData"
        description: "Default Username"
        required: false
        type: "string"
      - name: "recorder[default_password]"
        in: "formData"
        description: "Default Password"
        required: false
        type: "string"
      - name: "recorder[jpeg_url]"
        in: "formData"
        description: "JPEG URL"
        required: false
        type: "string"
      - name: "recorder[h264_url]"
        in: "formData"
        description: "H264 URL"
        required: false
        type: "string"
      - name: "recorder[mjpeg_url]"
        in: "formData"
        description: "MJPEG URL"
        required: false
        type: "string"
      - name: "recorder[official_url]"
        in: "formData"
        description: "Official URL"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Recorders"
    patch:
      tags:
      - "recorders"
      summary: "Updates an existing Recorder"
      operationId: "apiV1RecordersId_jsonPATCH"
      parameters:
      - name: "id"
        in: "path"
        description: "Recorder ID"
        required: true
        type: "string"
      - name: "vendor_id"
        in: "formData"
        description: "Vendor ID"
        required: true
        type: "string"
      - name: "recorder[model]"
        in: "formData"
        description: "Model"
        required: true
        type: "string"
      - name: "recorder[name]"
        in: "formData"
        description: "Name"
        required: true
        type: "string"
      - name: "recorder[recorder_type]"
        in: "formData"
        description: "Type"
        required: true
        type: "string"
      - name: "recorder[resolution]"
        in: "formData"
        description: "Resolution"
        required: false
        type: "string"
      - name: "recorder[onvif]"
        in: "formData"
        description: "ONVIF"
        required: false
        type: "string"
      - name: "recorder[psia]"
        in: "formData"
        description: "PSIA"
        required: false
        type: "string"
      - name: "recorder[ptz]"
        in: "formData"
        description: "PTZ"
        required: false
        type: "string"
      - name: "recorder[discontinued]"
        in: "formData"
        description: "Discontinued"
        required: false
        type: "string"
      - name: "recorder[support_3rdparty]"
        in: "formData"
        description: "3rd pparty Camera Support"
        required: false
        type: "string"
      - name: "recorder[sd_card]"
        in: "formData"
        description: "SD Card"
        required: false
        type: "string"
      - name: "recorder[upnp]"
        in: "formData"
        description: "UPnP"
        required: false
        type: "string"
      - name: "recorder[hot_swap]"
        in: "formData"
        description: "Hot Swap"
        required: false
        type: "string"
      - name: "recorder[hdmi]"
        in: "formData"
        description: "HDMI Support"
        required: false
        type: "string"
      - name: "recorder[digital_io]"
        in: "formData"
        description: "Digital I/O"
        required: false
        type: "string"
      - name: "recorder[audio_in]"
        in: "formData"
        description: "Audio In"
        required: false
        type: "string"
      - name: "recorder[audio_out]"
        in: "formData"
        description: "Audio Out"
        required: false
        type: "string"
      - name: "recorder[input_channels]"
        in: "formData"
        description: "Input Channels"
        required: false
        type: "string"
      - name: "recorder[playback_channels]"
        in: "formData"
        description: "Playback Channels"
        required: false
        type: "string"
      - name: "recorder[usb]"
        in: "formData"
        description: "USB Ports"
        required: false
        type: "string"
      - name: "recorder[sdhc]"
        in: "formData"
        description: "SD Card (GB)"
        required: false
        type: "string"
      - name: "recorder[mobile_access]"
        in: "formData"
        description: "Mobile Access"
        required: false
        type: "string"
      - name: "recorder[alarms]"
        in: "formData"
        description: "Alarms"
        required: false
        type: "string"
      - name: "recorder[raid_support]"
        in: "formData"
        description: "Raid Support"
        required: false
        type: "string"
      - name: "recorder[storage]"
        in: "formData"
        description: "Internal Storage"
        required: false
        type: "string"
      - name: "recorder[additional_information]"
        in: "formData"
        description: "Additional Information"
        required: false
        type: "string"
      - name: "recorder[default_username]"
        in: "formData"
        description: "Default Username"
        required: false
        type: "string"
      - name: "recorder[default_password]"
        in: "formData"
        description: "Default Password"
        required: false
        type: "string"
      - name: "recorder[jpeg_url]"
        in: "formData"
        description: "JPEG URL"
        required: false
        type: "string"
      - name: "recorder[h264_url]"
        in: "formData"
        description: "H264 URL"
        required: false
        type: "string"
      - name: "recorder[mjpeg_url]"
        in: "formData"
        description: "MJPEG URL"
        required: false
        type: "string"
      - name: "recorder[official_url]"
        in: "formData"
        description: "Official URL"
        required: false
        type: "string"
      responses:
        "200":
          description: "No response was specified"
        "401":
          description: "Unauthorized"
        "404":
          description: "Not Found"
        "406":
          description: "Not Acceptable"
      x-swagger-router-controller: "Recorders"
...